서버 전송 기준 수정
틱단위의 전송에서 시간에 의한 전송으로 변경

+++ 3d 게임/3Dproject/3Dproject/CGameFramework.h
@@ -187,7 +187,7 @@ public:
 	Color					m_color = blue;
 	std::string				m_NickName = "seojin";								//PlayName
 
-	int ServerFrameRate = 0;
+	float ServerFrameRate = 0;
 
 	bool	m_conneted = false;
 	//Recv정보 처리 함수
